+++ router/sr_icmp_handler.h
@@ -1,7 +1,7 @@
 #ifndef SR_ICMP_HANDLER_H
 #define SR_ICMP_HANDLER_H
 
-uint8_t* gen_icmp_packet (int type, int code);
+uint8_t* gen_icmp_packet (uint8_t type, uint8_t code);
 
 uint8_t* gen_eth_frame (uint8_t *packet, uint8_t *icmp_pkt, struct sr_if *interface);
